done inserting job applicant
but theres a bug. heres the scenario: if i submit using test@gmail.com, it would work but the next submit using the same email wouldnt respond with an error but with the third submit, it will now response with an error.

in conclusion, its like retiaining the previous value even if i already changed it from the second test