also made changes in the request leave dashboard
made all the decrementing functions for the leave remaining right based in the leave type id.

NOTE:
the decrementing function mustnt decrement the leave remaining immediately, it only must decrement it once approved.

decrement it based in the employee's intended days and add validation if the intended days is no longer valid if the remaining leave days is not sufficient